geshi use and regex fixes
regex for code splits

the change from explode to preg_split allows us to ignore inline code markup:

'Then you type ```ls -al``` to see...'

I also couldn't get ^ to target begining of the string (it prevented finding anything although it is correct syntax). So I eliminated it. But that made a simple '^' flag very risky (it could appear in text).

SO, I change the flag to '^&' which is far less likely